+++ geode-core/src/main/java/org/apache/geode/cache/configuration/DeclarableType.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.geode.internal.cache.configuration;
+package org.apache.geode.cache.configuration;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -26,24 +26,27 @@ import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlType;
 
+import org.apache.geode.annotations.Experimental;
+
 
 /**
  *
- * A "disk-dirs" element specifies the region's disk directories.
+ * A "declarable" element specifies a Declarable object to be placed in a Region entry.
  *
  *
  * <p>
- * Java class for disk-dirs-type complex type.
+ * Java class for declarable-type complex type.
  *
  * <p>
  * The following schema fragment specifies the expected content contained within this class.
  *
  * <pre>
- * &lt;complexType name="disk-dirs-type">
+ * &lt;complexType name="declarable-type">
  *   &lt;complexContent>
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *       &lt;sequence>
- *         &lt;element name="disk-dir" type="{http://geode.apache.org/schema/cache}disk-dir-type" maxOccurs="unbounded"/>
+ *         &lt;element name="class-name" type="{http://geode.apache.org/schema/cache}class-name-type"/>
+ *         &lt;element name="parameter" type="{http://geode.apache.org/schema/cache}parameter-type" maxOccurs="unbounded" minOccurs="0"/>
  *       &lt;/sequence>
  *     &lt;/restriction>
  *   &lt;/complexContent>
@@ -53,42 +56,69 @@ import javax.xml.bind.annotation.XmlType;
  *
  */
 @XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "disk-dirs-type", namespace = "http://geode.apache.org/schema/cache",
-    propOrder = {"diskDir"})
-public class DiskDirsType {
+@XmlType(name = "declarable-type", namespace = "http://geode.apache.org/schema/cache",
+    propOrder = {"className", "parameter"})
+@Experimental
+public class DeclarableType {
 
-  @XmlElement(name = "disk-dir", namespace = "http://geode.apache.org/schema/cache",
+  @XmlElement(name = "class-name", namespace = "http://geode.apache.org/schema/cache",
       required = true)
-  protected List<DiskDirType> diskDir;
+  protected String className;
+  @XmlElement(namespace = "http://geode.apache.org/schema/cache")
+  protected List<ParameterType> parameter;
 
   /**
-   * Gets the value of the diskDir property.
+   * Gets the value of the className property.
+   *
+   * @return
+   *         possible object is
+   *         {@link String }
+   *
+   */
+  public String getClassName() {
+    return className;
+  }
+
+  /**
+   * Sets the value of the className property.
+   *
+   * @param value
+   *        allowed object is
+   *        {@link String }
+   *
+   */
+  public void setClassName(String value) {
+    this.className = value;
+  }
+
+  /**
+   * Gets the value of the parameter property.
    *
    * <p>
    * This accessor method returns a reference to the live list,
    * not a snapshot. Therefore any modification you make to the
    * returned list will be present inside the JAXB object.
-   * This is why there is not a <CODE>set</CODE> method for the diskDir property.
+   * This is why there is not a <CODE>set</CODE> method for the parameter property.
    *
    * <p>
    * For example, to add a new item, do as follows:
    *
    * <pre>
-   * getDiskDir().add(newItem);
+   * getParameter().add(newItem);
    * </pre>
    *
    *
    * <p>
    * Objects of the following type(s) are allowed in the list
-   * {@link DiskDirType }
+   * {@link ParameterType }
    *
    *
    */
-  public List<DiskDirType> getDiskDir() {
-    if (diskDir == null) {
-      diskDir = new ArrayList<DiskDirType>();
+  public List<ParameterType> getParameter() {
+    if (parameter == null) {
+      parameter = new ArrayList<ParameterType>();
     }
-    return this.diskDir;
+    return this.parameter;
   }
 
 }
