+++ geode-core/src/main/java/org/apache/geode/cache/configuration/DiskDirType.java
@@ -16,13 +16,15 @@
  * limitations under the License.
  */
 
-package org.apache.geode.internal.cache.configuration;
+package org.apache.geode.cache.configuration;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlAttribute;
 import javax.xml.bind.annotation.XmlType;
 import javax.xml.bind.annotation.XmlValue;
+
+import org.apache.geode.annotations.Experimental;
 
 
 /**
@@ -51,6 +53,7 @@ import javax.xml.bind.annotation.XmlValue;
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "disk-dir-type", namespace = "http://geode.apache.org/schema/cache",
     propOrder = {"content"})
+@Experimental
 public class DiskDirType {
 
   @XmlValue
