+++ geode-core/src/main/java/org/apache/geode/cache/configuration/DiskDirsType.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.geode.internal.cache.configuration;
+package org.apache.geode.cache.configuration;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -26,27 +26,26 @@ import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlType;
 
+import org.apache.geode.annotations.Experimental;
+
 
 /**
  *
- * A "initializer" element specifies a Declarable object whose init method
- * will be called after all other cache.xml initialization is
- * complete.
+ * A "disk-dirs" element specifies the region's disk directories.
  *
  *
  * <p>
- * Java class for initializer-type complex type.
+ * Java class for disk-dirs-type complex type.
  *
  * <p>
  * The following schema fragment specifies the expected content contained within this class.
  *
  * <pre>
- * &lt;complexType name="initializer-type">
+ * &lt;complexType name="disk-dirs-type">
  *   &lt;complexContent>
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *       &lt;sequence>
- *         &lt;element name="class-name" type="{http://geode.apache.org/schema/cache}class-name-type"/>
- *         &lt;element name="parameter" type="{http://geode.apache.org/schema/cache}parameter-type" maxOccurs="unbounded" minOccurs="0"/>
+ *         &lt;element name="disk-dir" type="{http://geode.apache.org/schema/cache}disk-dir-type" maxOccurs="unbounded"/>
  *       &lt;/sequence>
  *     &lt;/restriction>
  *   &lt;/complexContent>
@@ -56,68 +55,43 @@ import javax.xml.bind.annotation.XmlType;
  *
  */
 @XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "initializer-type", namespace = "http://geode.apache.org/schema/cache",
-    propOrder = {"className", "parameter"})
-public class InitializerType {
+@XmlType(name = "disk-dirs-type", namespace = "http://geode.apache.org/schema/cache",
+    propOrder = {"diskDir"})
+@Experimental
+public class DiskDirsType {
 
-  @XmlElement(name = "class-name", namespace = "http://geode.apache.org/schema/cache",
+  @XmlElement(name = "disk-dir", namespace = "http://geode.apache.org/schema/cache",
       required = true)
-  protected String className;
-  @XmlElement(namespace = "http://geode.apache.org/schema/cache")
-  protected List<ParameterType> parameter;
+  protected List<DiskDirType> diskDir;
 
   /**
-   * Gets the value of the className property.
-   *
-   * @return
-   *         possible object is
-   *         {@link String }
-   *
-   */
-  public String getClassName() {
-    return className;
-  }
-
-  /**
-   * Sets the value of the className property.
-   *
-   * @param value
-   *        allowed object is
-   *        {@link String }
-   *
-   */
-  public void setClassName(String value) {
-    this.className = value;
-  }
-
-  /**
-   * Gets the value of the parameter property.
+   * Gets the value of the diskDir property.
    *
    * <p>
    * This accessor method returns a reference to the live list,
    * not a snapshot. Therefore any modification you make to the
    * returned list will be present inside the JAXB object.
-   * This is why there is not a <CODE>set</CODE> method for the parameter property.
+   * This is why there is not a <CODE>set</CODE> method for the diskDir property.
    *
    * <p>
    * For example, to add a new item, do as follows:
    *
    * <pre>
-   * getParameter().add(newItem);
+   * getDiskDir().add(newItem);
    * </pre>
    *
    *
    * <p>
    * Objects of the following type(s) are allowed in the list
-   * {@link ParameterType }
+   * {@link DiskDirType }
    *
    *
    */
-  public List<ParameterType> getParameter() {
-    if (parameter == null) {
-      parameter = new ArrayList<ParameterType>();
+  public List<DiskDirType> getDiskDir() {
+    if (diskDir == null) {
+      diskDir = new ArrayList<DiskDirType>();
     }
-    return this.parameter;
+    return this.diskDir;
   }
 
 }
