+++ geode-core/src/main/java/org/apache/geode/cache/configuration/EnumActionDestroyOverflow.java
@@ -16,43 +16,44 @@
  * limitations under the License.
  */
 
-package org.apache.geode.internal.cache.configuration;
+package org.apache.geode.cache.configuration;
 
 import javax.xml.bind.annotation.XmlEnum;
 import javax.xml.bind.annotation.XmlEnumValue;
 import javax.xml.bind.annotation.XmlType;
 
+import org.apache.geode.annotations.Experimental;
+
 
 /**
  * <p>
- * Java class for region-attributesMirror-type.
+ * Java class for enum-action-destroy-overflow.
  *
  * <p>
  * The following schema fragment specifies the expected content contained within this class.
  * <p>
  *
  * <pre>
- * &lt;simpleType name="region-attributesMirror-type">
+ * &lt;simpleType name="enum-action-destroy-overflow">
  *   &lt;restriction base="{http://www.w3.org/2001/XMLSchema}string">
- *     &lt;enumeration value="keys"/>
- *     &lt;enumeration value="keys-values"/>
- *     &lt;enumeration value="none"/>
+ *     &lt;enumeration value="local-destroy"/>
+ *     &lt;enumeration value="overflow-to-disk"/>
  *   &lt;/restriction>
  * &lt;/simpleType>
  * </pre>
  *
  */
-@XmlType(name = "region-attributesMirror-type", namespace = "http://geode.apache.org/schema/cache")
+@XmlType(name = "enum-action-destroy-overflow", namespace = "http://geode.apache.org/schema/cache")
 @XmlEnum
-public enum RegionAttributesMirrorType {
+@Experimental
+public enum EnumActionDestroyOverflow {
 
-  @XmlEnumValue("keys")
-  KEYS("keys"), @XmlEnumValue("keys-values")
-  KEYS_VALUES("keys-values"), @XmlEnumValue("none")
-  NONE("none");
+  @XmlEnumValue("local-destroy")
+  LOCAL_DESTROY("local-destroy"), @XmlEnumValue("overflow-to-disk")
+  OVERFLOW_TO_DISK("overflow-to-disk");
   private final String value;
 
-  RegionAttributesMirrorType(String v) {
+  EnumActionDestroyOverflow(String v) {
     value = v;
   }
 
@@ -60,8 +61,8 @@ public enum RegionAttributesMirrorType {
     return value;
   }
 
-  public static RegionAttributesMirrorType fromValue(String v) {
-    for (RegionAttributesMirrorType c : RegionAttributesMirrorType.values()) {
+  public static EnumActionDestroyOverflow fromValue(String v) {
+    for (EnumActionDestroyOverflow c : EnumActionDestroyOverflow.values()) {
       if (c.value.equals(v)) {
         return c;
       }
