+++ geode-core/src/main/java/org/apache/geode/cache/configuration/EnumReadableWritable.java
@@ -16,11 +16,13 @@
  * limitations under the License.
  */
 
-package org.apache.geode.internal.cache.configuration;
+package org.apache.geode.cache.configuration;
 
 import javax.xml.bind.annotation.XmlEnum;
 import javax.xml.bind.annotation.XmlEnumValue;
 import javax.xml.bind.annotation.XmlType;
+
+import org.apache.geode.annotations.Experimental;
 
 
 /**
@@ -43,6 +45,7 @@ import javax.xml.bind.annotation.XmlType;
  */
 @XmlType(name = "enum-readable-writable", namespace = "http://geode.apache.org/schema/cache")
 @XmlEnum
+@Experimental
 public enum EnumReadableWritable {
 
   @XmlEnumValue("readable")
