+++ geode-core/src/main/java/org/apache/geode/cache/configuration/ExpirationAttributesType.java
@@ -16,35 +16,37 @@
  * limitations under the License.
  */
 
-package org.apache.geode.internal.cache.configuration;
+package org.apache.geode.cache.configuration;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlAttribute;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlType;
+
+import org.apache.geode.annotations.Experimental;
 
 
 /**
  *
- * The "function service" element configures the behavior of the function service.
- * The function service allows users to execute functions on data stored in Geode.
+ * An "expiration-attributes" element describes expiration.
  *
  *
  * <p>
- * Java class for function-service-type complex type.
+ * Java class for expiration-attributes-type complex type.
  *
  * <p>
  * The following schema fragment specifies the expected content contained within this class.
  *
  * <pre>
- * &lt;complexType name="function-service-type">
+ * &lt;complexType name="expiration-attributes-type">
  *   &lt;complexContent>
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *       &lt;sequence>
- *         &lt;element name="function" maxOccurs="unbounded" minOccurs="0">
+ *         &lt;element name="custom-expiry" minOccurs="0">
  *           &lt;complexType>
  *             &lt;complexContent>
  *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
@@ -57,6 +59,17 @@ import javax.xml.bind.annotation.XmlType;
  *           &lt;/complexType>
  *         &lt;/element>
  *       &lt;/sequence>
+ *       &lt;attribute name="action">
+ *         &lt;simpleType>
+ *           &lt;restriction base="{http://www.w3.org/2001/XMLSchema}string">
+ *             &lt;enumeration value="destroy"/>
+ *             &lt;enumeration value="invalidate"/>
+ *             &lt;enumeration value="local-destroy"/>
+ *             &lt;enumeration value="local-invalidate"/>
+ *           &lt;/restriction>
+ *         &lt;/simpleType>
+ *       &lt;/attribute>
+ *       &lt;attribute name="timeout" use="required" type="{http://www.w3.org/2001/XMLSchema}string" />
  *     &lt;/restriction>
  *   &lt;/complexContent>
  * &lt;/complexType>
@@ -65,41 +78,88 @@ import javax.xml.bind.annotation.XmlType;
  *
  */
 @XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "function-service-type", namespace = "http://geode.apache.org/schema/cache",
-    propOrder = {"function"})
-public class FunctionServiceType {
+@XmlType(name = "expiration-attributes-type", namespace = "http://geode.apache.org/schema/cache",
+    propOrder = {"customExpiry"})
+@Experimental
+public class ExpirationAttributesType {
 
-  @XmlElement(namespace = "http://geode.apache.org/schema/cache")
-  protected List<FunctionServiceType.Function> function;
+  @XmlElement(name = "custom-expiry", namespace = "http://geode.apache.org/schema/cache")
+  protected ExpirationAttributesType.CustomExpiry customExpiry;
+  @XmlAttribute(name = "action")
+  protected String action;
+  @XmlAttribute(name = "timeout", required = true)
+  protected String timeout;
 
   /**
-   * Gets the value of the function property.
+   * Gets the value of the customExpiry property.
    *
-   * <p>
-   * This accessor method returns a reference to the live list,
-   * not a snapshot. Therefore any modification you make to the
-   * returned list will be present inside the JAXB object.
-   * This is why there is not a <CODE>set</CODE> method for the function property.
-   *
-   * <p>
-   * For example, to add a new item, do as follows:
-   *
-   * <pre>
-   * getFunction().add(newItem);
-   * </pre>
-   *
-   *
-   * <p>
-   * Objects of the following type(s) are allowed in the list
-   * {@link FunctionServiceType.Function }
-   *
+   * @return
+   *         possible object is
+   *         {@link ExpirationAttributesType.CustomExpiry }
    *
    */
-  public List<FunctionServiceType.Function> getFunction() {
-    if (function == null) {
-      function = new ArrayList<FunctionServiceType.Function>();
-    }
-    return this.function;
+  public ExpirationAttributesType.CustomExpiry getCustomExpiry() {
+    return customExpiry;
+  }
+
+  /**
+   * Sets the value of the customExpiry property.
+   *
+   * @param value
+   *        allowed object is
+   *        {@link ExpirationAttributesType.CustomExpiry }
+   *
+   */
+  public void setCustomExpiry(ExpirationAttributesType.CustomExpiry value) {
+    this.customExpiry = value;
+  }
+
+  /**
+   * Gets the value of the action property.
+   *
+   * @return
+   *         possible object is
+   *         {@link String }
+   *
+   */
+  public String getAction() {
+    return action;
+  }
+
+  /**
+   * Sets the value of the action property.
+   *
+   * @param value
+   *        allowed object is
+   *        {@link String }
+   *
+   */
+  public void setAction(String value) {
+    this.action = value;
+  }
+
+  /**
+   * Gets the value of the timeout property.
+   *
+   * @return
+   *         possible object is
+   *         {@link String }
+   *
+   */
+  public String getTimeout() {
+    return timeout;
+  }
+
+  /**
+   * Sets the value of the timeout property.
+   *
+   * @param value
+   *        allowed object is
+   *        {@link String }
+   *
+   */
+  public void setTimeout(String value) {
+    this.timeout = value;
   }
 
 
@@ -127,7 +187,7 @@ public class FunctionServiceType {
    */
   @XmlAccessorType(XmlAccessType.FIELD)
   @XmlType(name = "", propOrder = {"className", "parameter"})
-  public static class Function {
+  public static class CustomExpiry {
 
     @XmlElement(name = "class-name", namespace = "http://geode.apache.org/schema/cache",
         required = true)
