+++ geode-core/src/main/java/org/apache/geode/cache/configuration/PoolType.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.geode.internal.cache.configuration;
+package org.apache.geode.cache.configuration;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -26,6 +26,8 @@ import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlAttribute;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlType;
+
+import org.apache.geode.annotations.Experimental;
 
 
 /**
@@ -96,6 +98,7 @@ import javax.xml.bind.annotation.XmlType;
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "pool-type", namespace = "http://geode.apache.org/schema/cache",
     propOrder = {"locator", "server"})
+@Experimental
 public class PoolType {
 
   @XmlElement(namespace = "http://geode.apache.org/schema/cache")
